database/requirements.txt
Flask==3.1.2
flask-mysqldb==2.0.0
flask-cors==6.0.1
mysql-connector-python==9.5.0
bcrypt==5.0.0
PyJWT==2.10.1
python-dotenv==1.2.1